Remove AbstractOrmMapper.java
