+++ src/main/java/org/nkjmlab/sorm4j/internal/mapping/LazyResultSetImpl.java
@@ -27,18 +27,17 @@ final class LazyResultSetImpl<T> implements LazyResultSet<T> {
   private static final Class<? extends Map> MAP_CLASS = LinkedHashMap.class;
 
   private final Class<T> objectClass;
-  private final AbstractOrmMapper ormMapper;
+  private final OrmMapperImpl ormMapper;
   private final ResultSet resultSet;
   private final PreparedStatement stmt;
 
   @SuppressWarnings("unchecked")
-  public LazyResultSetImpl(AbstractOrmMapper ormMapper, PreparedStatement stmt,
-      ResultSet resultSet) {
+  public LazyResultSetImpl(OrmMapperImpl ormMapper, PreparedStatement stmt, ResultSet resultSet) {
     this(ormMapper, (Class<T>) MAP_CLASS, stmt, resultSet);
   }
 
-  public LazyResultSetImpl(AbstractOrmMapper ormMapper, Class<T> objectClass,
-      PreparedStatement stmt, ResultSet resultSet) {
+  public LazyResultSetImpl(OrmMapperImpl ormMapper, Class<T> objectClass, PreparedStatement stmt,
+      ResultSet resultSet) {
     this.ormMapper = ormMapper;
     this.objectClass = objectClass;
     this.stmt = stmt;
@@ -82,10 +81,9 @@ final class LazyResultSetImpl<T> implements LazyResultSet<T> {
   @Override
   public List<T> toList() {
     @SuppressWarnings("unchecked")
-    List<T> ret = Try.getOrThrow(
-        () -> objectClass.equals(MAP_CLASS) ? (List<T>) ormMapper.mapRowsToMapList(resultSet)
-            : ormMapper.loadPojoList(objectClass, resultSet),
-        Try::rethrow);
+    List<T> ret = Try.getOrThrow(() -> objectClass.equals(MAP_CLASS)
+        ? (List<T>) Try.getOrThrow(() -> ormMapper.mapRowsAux(resultSet), Try::rethrow)
+        : ormMapper.loadPojoList(objectClass, resultSet), Try::rethrow);
     close();
     return ret;
   }
@@ -140,11 +138,11 @@ final class LazyResultSetImpl<T> implements LazyResultSet<T> {
     private final Supplier<S> getFunction;
 
     @SuppressWarnings("unchecked")
-    public LazyResultSetIterator(AbstractOrmMapper orMapper, Class<S> objectClass,
+    public LazyResultSetIterator(OrmMapperImpl ormMapper, Class<S> objectClass,
         PreparedStatement stmt, ResultSet resultSet) {
       this.getFunction = objectClass.equals(MAP_CLASS)
-          ? Try.createSupplierWithThrow(() -> (S) orMapper.mapRowAux(resultSet), Try::rethrow)
-          : Try.createSupplierWithThrow(() -> orMapper.mapRowAux(objectClass, resultSet),
+          ? Try.createSupplierWithThrow(() -> (S) ormMapper.mapRowAux(resultSet), Try::rethrow)
+          : Try.createSupplierWithThrow(() -> ormMapper.mapRowAux(objectClass, resultSet),
               Try::rethrow);
     }
