+++ src/main/java/org/nkjmlab/sorm4j/internal/mapping/Mappings.java
@@ -23,6 +23,7 @@ import org.nkjmlab.sorm4j.internal.util.StringUtils;
 import org.nkjmlab.sorm4j.internal.util.Try;
 
 class Mappings {
+  static final org.slf4j.Logger log = org.nkjmlab.sorm4j.internal.util.LoggerFactory.getLogger();
 
   private final TableNameMapper tableNameMapper;
   private final ColumnFieldMapper columnFieldMapper;
@@ -79,7 +80,7 @@ class Mappings {
         (TableMapping<T>) tableMappings.computeIfAbsent(key, Try.createFunctionWithThrow(_key -> {
           TableMapping<T> m = createTableMapping(objectClass, tableName.getName(), connection);
           LogPointFactory.createLogPoint(SormLogger.Category.MAPPING)
-              .ifPresent(lp -> AbstractOrmMapper.log.info("[{}]" + System.lineSeparator() + "{}",
+              .ifPresent(lp -> log.info("[{}]" + System.lineSeparator() + "{}",
                   lp.getTag(), m.getFormattedString()));
           return m;
         }, Try::rethrow));
@@ -148,7 +149,7 @@ class Mappings {
       ColumnsMapping<T> m = createColumnsMapping(objectClass);
 
       LogPointFactory.createLogPoint(SormLogger.Category.MAPPING).ifPresent(
-          lp -> AbstractOrmMapper.log.info(System.lineSeparator() + m.getFormattedString()));
+          lp -> log.info(System.lineSeparator() + m.getFormattedString()));
 
       return m;
     });
