+++ src/main/java/org/nkjmlab/sorm4j/internal/mapping/OrmMapperImpl.java
@@ -1,93 +1,133 @@
 
 package org.nkjmlab.sorm4j.internal.mapping;
 
+import static org.nkjmlab.sorm4j.internal.util.StringUtils.*;
 import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.ResultSetMetaData;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.function.Function;
+import java.util.function.Supplier;
+import org.nkjmlab.sorm4j.FunctionHandler;
+import org.nkjmlab.sorm4j.RowMapper;
+import org.nkjmlab.sorm4j.SormException;
+import org.nkjmlab.sorm4j.SormLogger;
+import org.nkjmlab.sorm4j.extension.ResultSetConverter;
+import org.nkjmlab.sorm4j.extension.SqlParameterSetter;
+import org.nkjmlab.sorm4j.internal.util.LogPoint;
+import org.nkjmlab.sorm4j.internal.util.LogPointFactory;
+import org.nkjmlab.sorm4j.internal.util.Try;
 import org.nkjmlab.sorm4j.sql.InsertResult;
 import org.nkjmlab.sorm4j.sql.LazyResultSet;
 import org.nkjmlab.sorm4j.sql.SqlStatement;
+import org.nkjmlab.sorm4j.sql.tuple.Tuple2;
+import org.nkjmlab.sorm4j.sql.tuple.Tuple3;
+import org.nkjmlab.sorm4j.sql.tuple.Tuples;
 
 /**
  * The main class for the ORMapper engine.
  *
  */
-class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
+class OrmMapperImpl implements OrmMapper {
+
+  static final org.slf4j.Logger log = org.nkjmlab.sorm4j.internal.util.LoggerFactory.getLogger();
+
+  @SuppressWarnings("unchecked")
+  private static <T, R> R applytoArray(List<T> objects, Function<T[], R> sqlFunc) {
+    return sqlFunc.apply((T[]) objects.toArray(Object[]::new));
+  }
+
+  static <R> R executeQueryAndRead(Connection connection, SqlParameterSetter sqlParameterSetter,
+      String sql, Object[] parameters, FunctionHandler<ResultSet, R> resultSetHandler) {
+    final Optional<LogPoint> dp = LogPointFactory.createLogPoint(SormLogger.Category.EXECUTE_QUERY);
+    dp.ifPresent(lp -> {
+      log.debug("[{}] [{}] with {} parameters", lp.getTag(), sql,
+          parameters == null ? 0 : parameters.length);
+      log.trace("[{}] Parameters = {}", lp.getTag(), parameters);
+    });
+
+    try (PreparedStatement stmt = connection.prepareStatement(sql)) {
+      sqlParameterSetter.setParameters(stmt, parameters);
+      try (ResultSet resultSet = stmt.executeQuery()) {
+        R ret = resultSetHandler.apply(resultSet);
+        dp.ifPresent(sw -> log.debug("{} Read [{}] objects from [{}]", sw.getTagAndElapsedTime(),
+            ret instanceof Collection ? ((Collection<?>) ret).size() : 1,
+            Try.getOrNull(() -> connection.getMetaData().getURL())));
+        return ret;
+      }
+    } catch (Exception e) {
+      String msg = (parameters == null || parameters.length == 0) ? format("Error in sql=[{}]", sql)
+          : format("Fail to execute sql=[{}], parameters={}", sql, parameters);
+      throw new SormException(msg + System.lineSeparator() + e.getMessage(), e);
+    }
+  }
+
+  static final int executeUpdateAndClose(Connection connection,
+      SqlParameterSetter sqlParameterSetter, String sql, Object[] parameters) {
+    try (PreparedStatement stmt = connection.prepareStatement(sql)) {
+      sqlParameterSetter.setParameters(stmt, parameters);
+      return stmt.executeUpdate();
+    } catch (Exception e) {
+      throw Try.rethrow(e);
+    }
+  }
+
+  private final Connection connection;
+
+  private final Mappings mappings;
+
+
+  private final ResultSetConverter resultSetConverter;
+
+  private final SqlParameterSetter sqlParameterSetter;
+
+
+  private final int transactionIsolationLevel;
+
+
+  private final List<LazyResultSet<?>> lazyResultSets = new ArrayList<>();
 
   /**
    * Creates a instance that will use the default cache for table-object and column-object mappings.
    *
    * @param connection {@link java.sql.Connection} object to be used
+   * @param configStore
    */
-
-  public OrmMapperImpl(Connection connection, ConfigStore defaultConfigurations) {
-    super(connection, defaultConfigurations);
-  }
-
-
-  @Override
-  public <T> int insert(T object) {
-    return getCastedTableMapping(object.getClass()).insert(getJdbcConnection(), object);
-  }
-
-  @Override
-  public <T> int insertOn(String tableName, T object) {
-    return getCastedTableMapping(tableName, object.getClass()).insert(getJdbcConnection(), object);
+  public OrmMapperImpl(Connection connection, ConfigStore configStore) {
+    this.connection = connection;
+    this.mappings = configStore.getMappings();
+    this.resultSetConverter = mappings.getResultSetConverter();
+    this.sqlParameterSetter = mappings.getSqlParameterSetter();
+    this.transactionIsolationLevel = configStore.getTransactionIsolationLevel();
   }
 
 
 
   @Override
-  public <T> InsertResult<T> insertAndGet(T object) {
-    TableMapping<T> mapping = getCastedTableMapping(object.getClass());
-    return mapping.insertAndGet(getJdbcConnection(), object);
+  public void close() {
+    Try.runOrThrow(() -> {
+      lazyResultSets.forEach(rs -> rs.close());
+      lazyResultSets.clear();
+      getJdbcConnection().close();
+    }, Try::rethrow);
   }
-
 
   @Override
-  public <T> InsertResult<T> insertAndGetOn(String tableName, T object) {
-    TableMapping<T> mapping = getCastedTableMapping(tableName, object.getClass());
-    return mapping.insertAndGet(getJdbcConnection(), object);
+  public <T> int[] delete(List<T> objects) {
+    return applytoArray(objects, array -> delete(array));
   }
-
 
 
   @Override
   public <T> int delete(T object) {
     return getCastedTableMapping(object.getClass()).delete(getJdbcConnection(), object);
   }
-
-
-  @Override
-  public <T> int deleteOn(String tableName, T object) {
-    return getCastedTableMapping(tableName, object.getClass()).delete(getJdbcConnection(), object);
-  }
-
-
-  @Override
-  public <T> int update(T object) {
-    return getCastedTableMapping(object.getClass()).update(getJdbcConnection(), object);
-  }
-
-
-  @Override
-  public <T> int updateOn(String tableName, T object) {
-    return getCastedTableMapping(tableName, object.getClass()).update(getJdbcConnection(), object);
-  }
-
-
-  @Override
-  public <T> int merge(T object) {
-    return getCastedTableMapping(object.getClass()).merge(getJdbcConnection(), object);
-  }
-
-
-  @Override
-  public <T> int mergeOn(String tableName, T object) {
-    return getCastedTableMapping(tableName, object.getClass()).merge(getJdbcConnection(), object);
-  }
-
 
   @Override
   public <T> int[] delete(@SuppressWarnings("unchecked") T... objects) {
@@ -97,9 +137,341 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
 
 
   @Override
-  public <T> int[] update(@SuppressWarnings("unchecked") T... objects) {
+  public <T> int deleteAll(Class<T> objectClass) {
+    return deleteAllOn(getTableName(objectClass));
+  }
+
+  @Override
+  public int deleteAllOn(String tableName) {
+    return executeUpdate("DELETE FROM " + tableName);
+  }
+
+
+  @Override
+  public <T> int[] deleteOn(String tableName, List<T> objects) {
+    return applytoArray(objects, array -> deleteOn(tableName, array));
+  }
+
+
+  @Override
+  public <T> int deleteOn(String tableName, T object) {
+    return getCastedTableMapping(tableName, object.getClass()).delete(getJdbcConnection(), object);
+  }
+
+  @Override
+  public <T> int[] deleteOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(tableName, objects,
+        mapping -> mapping.delete(getJdbcConnection(), objects), () -> new int[0]);
+  }
+
+
+  /**
+   * Execute sql function with table name. objects when objects[0] is null,
+   * {@code NullPointerException} are throw.
+   */
+  private final <T, R> R execSqlIfParameterExists(String tableName, T[] objects,
+      Function<TableMapping<T>, R> sqlFunction, Supplier<R> notExists) {
+    if (objects == null || objects.length == 0) {
+      return notExists.get();
+    }
+    TableMapping<T> mapping = getCastedTableMapping(tableName, objects[0].getClass());
+    return sqlFunction.apply(mapping);
+  }
+
+
+  /**
+   * Execute sql function. objects when objects[0] is null, {@code NullPointerException} are throw.
+   */
+  private final <T, R> R execSqlIfParameterExists(T[] objects,
+      Function<TableMapping<T>, R> sqlFunction, Supplier<R> notExists) {
+    if (objects == null || objects.length == 0) {
+      return notExists.get();
+    }
+    TableMapping<T> mapping = getCastedTableMapping(objects[0].getClass());
+    return sqlFunction.apply(mapping);
+  }
+
+  @Override
+  public <T> T executeQuery(SqlStatement sql, FunctionHandler<ResultSet, T> resultSetHandler) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql.getSql(),
+        sql.getParameters(), resultSetHandler);
+  }
+
+  @Override
+  public <T> List<T> executeQuery(SqlStatement sql, RowMapper<T> rowMapper) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql.getSql(),
+        sql.getParameters(), RowMapper.convertToRowsMapper(rowMapper));
+  }
+
+  @Override
+  public int executeUpdate(SqlStatement sql) {
+    return executeUpdate(sql.getSql(), sql.getParameters());
+  }
+
+  @Override
+  public int executeUpdate(String sql, Object... parameters) {
+    final Optional<LogPoint> dp =
+        LogPointFactory.createLogPoint(SormLogger.Category.EXECUTE_UPDATE);
+
+    final int ret = executeUpdateAndClose(connection, sqlParameterSetter, sql, parameters);
+    dp.ifPresent(sw -> {
+      log.trace("[{}] Parameters = {} ", sw.getTag(), parameters);
+      log.debug("{} Call [{}] [{}]", sw.getTagAndElapsedTime(), sql,
+          Try.getOrNull(() -> connection.getMetaData().getURL()), sql);
+    });
+    return ret;
+  }
+
+
+  private <T> T getAux(Class<T> objectClass, ResultSet resultSet) {
+    final ColumnsMapping<T> m = getColumnsMapping(objectClass);
+    return Try.getOrThrow(() -> m.loadPojo(m.createColumnLabels(resultSet), resultSet),
+        Try::rethrow);
+  }
+
+  private <T> TableMapping<T> getCastedTableMapping(Class<?> objectClass) {
+    return mappings.getCastedTableMapping(connection, objectClass);
+  }
+
+  private <T> TableMapping<T> getCastedTableMapping(String tableName, Class<?> objectClass) {
+    return mappings.getCastedTableMapping(connection, tableName, objectClass);
+  }
+
+  <T> ColumnsMapping<T> getColumnsMapping(Class<T> objectClass) {
+    return mappings.getColumnsMapping(objectClass);
+  }
+
+  @Override
+  public Connection getJdbcConnection() {
+    return connection;
+  }
+
+  <T> TableMapping<T> getTableMapping(Class<T> objectClass) {
+    return mappings.getTableMapping(connection, objectClass);
+  }
+
+  @Override
+  public String getTableName(Class<?> objectClass) {
+    return mappings.getTableName(connection, objectClass);
+  }
+
+  private int getTransactionIsolationLevel() {
+    return transactionIsolationLevel;
+  }
+
+  @Override
+  public <T> int[] insert(List<T> objects) {
+    return applytoArray(objects, array -> insert(array));
+  }
+
+  @Override
+  public <T> int insert(T object) {
+    return getCastedTableMapping(object.getClass()).insert(getJdbcConnection(), object);
+  }
+
+  @Override
+  public <T> int[] insert(@SuppressWarnings("unchecked") T... objects) {
     return execSqlIfParameterExists(objects,
-        mapping -> mapping.update(getJdbcConnection(), objects), () -> new int[0]);
+        mapping -> mapping.insert(getJdbcConnection(), objects), () -> new int[0]);
+  }
+
+  @Override
+  public <T> InsertResult<T> insertAndGet(List<T> objects) {
+    return applytoArray(objects, array -> insertAndGet(array));
+  }
+
+
+  @Override
+  public <T> InsertResult<T> insertAndGet(T object) {
+    TableMapping<T> mapping = getCastedTableMapping(object.getClass());
+    return mapping.insertAndGet(getJdbcConnection(), object);
+  }
+
+  @Override
+  public <T> InsertResult<T> insertAndGet(@SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(objects,
+        mapping -> mapping.insertAndGet(getJdbcConnection(), objects),
+        () -> InsertResultImpl.emptyInsertResult());
+  }
+
+
+  @Override
+  public <T> InsertResult<T> insertAndGetOn(String tableName, List<T> objects) {
+    return applytoArray(objects, array -> insertAndGetOn(tableName, array));
+  }
+
+  @Override
+  public <T> InsertResult<T> insertAndGetOn(String tableName, T object) {
+    TableMapping<T> mapping = getCastedTableMapping(tableName, object.getClass());
+    return mapping.insertAndGet(getJdbcConnection(), object);
+  }
+
+  @Override
+  public <T> InsertResult<T> insertAndGetOn(String tableName,
+      @SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(tableName, objects,
+        mapping -> mapping.insertAndGet(getJdbcConnection(), objects),
+        () -> InsertResultImpl.emptyInsertResult());
+  }
+
+  @Override
+  public <T> int[] insertOn(String tableName, List<T> objects) {
+    return applytoArray(objects, array -> insertOn(tableName, array));
+  }
+
+
+  @Override
+  public <T> int insertOn(String tableName, T object) {
+    return getCastedTableMapping(tableName, object.getClass()).insert(getJdbcConnection(), object);
+  }
+
+
+  @Override
+  public <T> int[] insertOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(tableName, objects,
+        mapping -> mapping.insert(getJdbcConnection(), objects), () -> new int[0]);
+  }
+
+  <T> T loadFirst(Class<T> objectClass, ResultSet resultSet) throws SQLException {
+    if (resultSet.next()) {
+      return mapRowAux(objectClass, resultSet);
+    }
+    return null;
+  }
+
+
+
+  Map<String, Object> loadFirstMap(ResultSet resultSet) throws SQLException {
+    Map<String, Object> ret = null;
+    if (resultSet.next()) {
+      ret = mapRowAux(resultSet);
+    }
+    return ret;
+  }
+
+
+  private final <T> List<T> loadNativeObjectList(Class<T> objectClass, ResultSet resultSet)
+      throws SQLException {
+    Try.runOrThrow(() -> {
+      ResultSetMetaData metaData = resultSet.getMetaData();
+      if (metaData.getColumnCount() != 1) {
+        throw new SormException("ResultSet returned [" + metaData.getColumnCount()
+            + "] columns but 1 column was expected to load data into an instance of ["
+            + objectClass.getName() + "]");
+      }
+    }, Try::rethrow);
+    final List<T> ret = new ArrayList<>();
+    while (resultSet.next()) {
+      ret.add(resultSetConverter.toSingleNativeObject(resultSet, objectClass));
+    }
+    return ret;
+
+  }
+
+
+
+  <T> T loadOne(Class<T> objectClass, ResultSet resultSet) throws SQLException {
+    T ret = null;
+    if (resultSet.next()) {
+      ret = mapRowAux(objectClass, resultSet);
+    }
+    if (resultSet.next()) {
+      throw new RuntimeException("Non-unique result returned");
+    }
+    return ret;
+  }
+
+
+  Map<String, Object> loadOneMap(ResultSet resultSet) throws SQLException {
+    Map<String, Object> ret = null;
+    if (resultSet.next()) {
+      ret = mapRowAux(resultSet);
+    }
+    if (resultSet.next()) {
+      throw new SormException("Non-unique result returned");
+    }
+    return ret;
+  }
+
+
+  public final <T> List<T> loadPojoList(final Class<T> objectClass, final ResultSet resultSet)
+      throws SQLException {
+    ColumnsMapping<T> mapping = mappings.getColumnsMapping(objectClass);
+    return mapping.loadPojoList(resultSet);
+  }
+
+
+  private final <T> T loadSinglePojo(final Class<T> objectClass, final ResultSet resultSet)
+      throws SQLException {
+    ColumnsMapping<T> mapping = mappings.getColumnsMapping(objectClass);
+    return mapping.loadPojo(resultSet);
+  }
+
+
+  @Override
+  public <T> T mapRow(Class<T> objectClass, ResultSet resultSet) {
+    return Try.getOrThrow(() -> mapRowAux(objectClass, resultSet), Try::rethrow);
+  }
+
+
+  <T> T mapRowAux(Class<T> objectClass, ResultSet resultSet) throws SQLException {
+    return resultSetConverter.isEnableToConvertNativeObject(objectClass)
+        ? resultSetConverter.toSingleNativeObject(resultSet, objectClass)
+        : loadSinglePojo(objectClass, resultSet);
+  }
+
+
+  Map<String, Object> mapRowAux(ResultSet resultSet) throws SQLException {
+    ColumnsAndTypes ct = ColumnsAndTypes.createColumnsAndTypes(resultSet);
+    return resultSetConverter.toSingleMap(resultSet, ct.getColumns(), ct.getColumnTypes());
+  }
+
+
+  @Override
+  public <T> List<T> mapRows(Class<T> objectClass, ResultSet resultSet) {
+    return Try.getOrThrow(() -> mapRowsAux(objectClass, resultSet), Try::rethrow);
+  }
+
+
+  final <T> List<T> mapRowsAux(Class<T> objectClass, ResultSet resultSet) throws SQLException {
+    return resultSetConverter.isEnableToConvertNativeObject(objectClass)
+        ? loadNativeObjectList(objectClass, resultSet)
+        : loadPojoList(objectClass, resultSet);
+  }
+
+
+  final List<Map<String, Object>> mapRowsAux(ResultSet resultSet) throws SQLException {
+    final List<Map<String, Object>> ret = new ArrayList<>();
+    ColumnsAndTypes ct = ColumnsAndTypes.createColumnsAndTypes(resultSet);
+    while (resultSet.next()) {
+      ret.add(resultSetConverter.toSingleMap(resultSet, ct.getColumns(), ct.getColumnTypes()));
+    }
+    return ret;
+  }
+
+
+  @Override
+  public List<Map<String, Object>> mapRowsToMapList(ResultSet resultSet) {
+    return Try.getOrThrow(() -> mapRowsAux(resultSet), Try::rethrow);
+  }
+
+
+
+  @Override
+  public Map<String, Object> mapRowToMap(ResultSet resultSet) {
+    return Try.getOrThrow(() -> mapRowAux(resultSet), Try::rethrow);
+  }
+
+
+  @Override
+  public <T> int[] merge(List<T> objects) {
+    return applytoArray(objects, array -> merge(array));
+  }
+
+
+  @Override
+  public <T> int merge(T object) {
+    return getCastedTableMapping(object.getClass()).merge(getJdbcConnection(), object);
   }
 
 
@@ -111,34 +483,16 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
 
 
   @Override
-  public <T> int[] insert(@SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(objects,
-        mapping -> mapping.insert(getJdbcConnection(), objects), () -> new int[0]);
-  }
-
-
-  @Override
-  public <T> InsertResult<T> insertAndGet(@SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(objects,
-        mapping -> mapping.insertAndGet(getJdbcConnection(), objects),
-        () -> InsertResultImpl.emptyInsertResult());
+  public <T> int[] mergeOn(String tableName, List<T> objects) {
+    return applytoArray(objects, array -> mergeOn(tableName, array));
   }
 
 
 
   @Override
-  public <T> int[] deleteOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(tableName, objects,
-        mapping -> mapping.delete(getJdbcConnection(), objects), () -> new int[0]);
+  public <T> int mergeOn(String tableName, T object) {
+    return getCastedTableMapping(tableName, object.getClass()).merge(getJdbcConnection(), object);
   }
-
-
-  @Override
-  public <T> int[] updateOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(tableName, objects,
-        mapping -> mapping.update(getJdbcConnection(), objects), () -> new int[0]);
-  }
-
 
   @Override
   public <T> int[] mergeOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
@@ -148,99 +502,15 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
 
 
   @Override
-  public <T> int[] insertOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(tableName, objects,
-        mapping -> mapping.insert(getJdbcConnection(), objects), () -> new int[0]);
-  }
-
-
-  @Override
-  public <T> InsertResult<T> insertAndGetOn(String tableName,
-      @SuppressWarnings("unchecked") T... objects) {
-    return execSqlIfParameterExists(tableName, objects,
-        mapping -> mapping.insertAndGet(getJdbcConnection(), objects),
-        () -> InsertResultImpl.emptyInsertResult());
-  }
-
-
-
-  @Override
-  public <T> int[] delete(List<T> objects) {
-    return applytoArray(objects, array -> delete(array));
-  }
-
-  @SuppressWarnings("unchecked")
-  private static <T, R> R applytoArray(List<T> objects, Function<T[], R> sqlFunc) {
-    return sqlFunc.apply((T[]) objects.toArray(Object[]::new));
-  }
-
-
-  @Override
-  public <T> int[] deleteOn(String tableName, List<T> objects) {
-    return applytoArray(objects, array -> deleteOn(tableName, array));
-  }
-
-  @Override
-  public <T> int[] insert(List<T> objects) {
-    return applytoArray(objects, array -> insert(array));
-  }
-
-
-  @Override
-  public <T> InsertResult<T> insertAndGet(List<T> objects) {
-    return applytoArray(objects, array -> insertAndGet(array));
-  }
-
-
-  @Override
-  public <T> InsertResult<T> insertAndGetOn(String tableName, List<T> objects) {
-    return applytoArray(objects, array -> insertAndGetOn(tableName, array));
-  }
-
-
-  @Override
-  public <T> int[] insertOn(String tableName, List<T> objects) {
-    return applytoArray(objects, array -> insertOn(tableName, array));
-  }
-
-
-  @Override
-  public <T> int[] merge(List<T> objects) {
-    return applytoArray(objects, array -> merge(array));
-  }
-
-
-  @Override
-  public <T> int[] mergeOn(String tableName, List<T> objects) {
-    return applytoArray(objects, array -> mergeOn(tableName, array));
-  }
-
-
-  @Override
-  public <T> int[] updateOn(String tableName, List<T> objects) {
-    return applytoArray(objects, array -> updateOn(tableName, array));
-  }
-
-
-  @Override
-  public <T> int[] update(List<T> objects) {
-    return applytoArray(objects, array -> update(array));
-  }
-
-  @Override
-  public <T> T readByPrimaryKey(Class<T> objectClass, Object... primaryKeyValues) {
-    return readByPrimaryKeyAux(objectClass, primaryKeyValues);
-  }
-
-  @Override
-  public <T> List<T> readList(Class<T> objectClass, String sql, Object... parameters) {
-    return readListAux(objectClass, sql, parameters);
-  }
-
-
-  @Override
   public final <T> List<T> readAll(Class<T> objectClass) {
     return readAllAux(objectClass);
+  }
+
+  /**
+   * Reads a list of all objects in the database mapped to the given object class.
+   */
+  final <T> List<T> readAllAux(final Class<T> objectClass) {
+    return readListAux(objectClass, getCastedTableMapping(objectClass).getSql().getSelectAllSql());
   }
 
 
@@ -249,14 +519,27 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
     return readAllLazyAux(objectClass);
   }
 
-  @Override
-  public <T> T readFirst(Class<T> objectClass, String sql, Object... parameters) {
-    return readFirstAux(objectClass, sql, parameters);
+
+  final <T> LazyResultSet<T> readAllLazyAux(Class<T> objectClass) {
+    return readLazyAux(objectClass,
+        mappings.getTableMapping(connection, objectClass).getSql().getSelectAllSql());
   }
 
+
   @Override
-  public <T> LazyResultSet<T> readLazy(Class<T> objectClass, String sql, Object... parameters) {
-    return readLazyAux(objectClass, sql, parameters);
+  public <T> T readByPrimaryKey(Class<T> objectClass, Object... primaryKeyValues) {
+    return readByPrimaryKeyAux(objectClass, primaryKeyValues);
+  }
+
+
+  /**
+   * Reads an object from the database by its primary keys.
+   */
+  final <T> T readByPrimaryKeyAux(final Class<T> objectClass, final Object... primaryKeyValues) {
+    final TableMapping<T> mapping = mappings.getTableMapping(connection, objectClass);
+    mapping.throwExeptionIfPrimaryKeysIsNotExist();
+    final String sql = mapping.getSql().getSelectByPrimaryKeySql();
+    return readFirstAux(objectClass, sql, primaryKeyValues);
   }
 
 
@@ -265,10 +548,42 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
     return readFirst(objectClass, sql.getSql(), sql.getParameters());
   }
 
+
+  @Override
+  public <T> T readFirst(Class<T> objectClass, String sql, Object... parameters) {
+    return readFirstAux(objectClass, sql, parameters);
+  }
+
+
+  final <T> T readFirstAux(Class<T> objectClass, String sql, Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> loadFirst(objectClass, resultSet));
+  }
+
   @Override
   public <T> LazyResultSet<T> readLazy(Class<T> objectClass, SqlStatement sql) {
     return readLazy(objectClass, sql.getSql(), sql.getParameters());
   }
+
+  @Override
+  public <T> LazyResultSet<T> readLazy(Class<T> objectClass, String sql, Object... parameters) {
+    return readLazyAux(objectClass, sql, parameters);
+  }
+
+
+  final <T> LazyResultSet<T> readLazyAux(Class<T> objectClass, String sql, Object... parameters) {
+    try {
+      final PreparedStatement stmt = connection.prepareStatement(sql);
+      sqlParameterSetter.setParameters(stmt, parameters);
+      final ResultSet resultSet = stmt.executeQuery();
+      LazyResultSetImpl<T> ret = new LazyResultSetImpl<T>(this, objectClass, stmt, resultSet);
+      lazyResultSets.add(ret);
+      return ret;
+    } catch (SQLException e) {
+      throw Try.rethrow(e);
+    }
+  }
+
 
   @Override
   public <T> List<T> readList(Class<T> objectClass, SqlStatement sql) {
@@ -276,13 +591,31 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
   }
 
   @Override
+  public <T> List<T> readList(Class<T> objectClass, String sql, Object... parameters) {
+    return readListAux(objectClass, sql, parameters);
+  }
+
+  final <T> List<T> readListAux(Class<T> objectClass, String sql, Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> mapRowsAux(objectClass, resultSet));
+  }
+
+
+  @Override
   public Map<String, Object> readMapFirst(SqlStatement sql) {
     return readMapFirst(sql.getSql(), sql.getParameters());
   }
 
   @Override
-  public Map<String, Object> readMapOne(SqlStatement sql) {
-    return readMapOne(sql.getSql(), sql.getParameters());
+  public Map<String, Object> readMapFirst(final String sql, final Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> {
+          ColumnsAndTypes ct = ColumnsAndTypes.createColumnsAndTypes(resultSet);
+          if (resultSet.next()) {
+            return resultSetConverter.toSingleMap(resultSet, ct.getColumns(), ct.getColumnTypes());
+          }
+          return null;
+        });
   }
 
   @Override
@@ -291,8 +624,55 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
   }
 
   @Override
+  public LazyResultSet<Map<String, Object>> readMapLazy(String sql, Object... parameters) {
+    try {
+      final PreparedStatement stmt = connection.prepareStatement(sql);
+      sqlParameterSetter.setParameters(stmt, parameters);
+      final ResultSet resultSet = stmt.executeQuery();
+      @SuppressWarnings({"unchecked", "rawtypes", "resource"})
+      LazyResultSet<Map<String, Object>> ret =
+          (LazyResultSet<Map<String, Object>>) new LazyResultSetImpl(this, stmt, resultSet);
+      return ret;
+    } catch (SQLException e) {
+      throw Try.rethrow(e);
+    }
+  }
+
+  @Override
   public List<Map<String, Object>> readMapList(SqlStatement sql) {
     return readMapList(sql.getSql(), sql.getParameters());
+  }
+
+  @Override
+  public List<Map<String, Object>> readMapList(final String sql, final Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> mapRowsAux(resultSet));
+  }
+
+  @Override
+  public Map<String, Object> readMapOne(SqlStatement sql) {
+    return readMapOne(sql.getSql(), sql.getParameters());
+  }
+
+  @Override
+  public Map<String, Object> readMapOne(final String sql, final Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> {
+          ColumnsAndTypes ct = ColumnsAndTypes.createColumnsAndTypes(resultSet);
+          Map<String, Object> ret = null;
+          if (resultSet.next()) {
+            ret = resultSetConverter.toSingleMap(resultSet, ct.getColumns(), ct.getColumnTypes());
+          }
+          if (resultSet.next()) {
+            throw new SormException("Non-unique result returned");
+          }
+          return ret;
+        });
+  }
+
+  @Override
+  public <T> T readOne(Class<T> objectClass, SqlStatement sql) {
+    return readOneAux(objectClass, sql.getSql(), sql.getParameters());
   }
 
   @Override
@@ -300,9 +680,125 @@ class OrmMapperImpl extends AbstractOrmMapper implements OrmMapper {
     return readOneAux(objectClass, sql, parameters);
   }
 
+
+
+  final <T> T readOneAux(final Class<T> objectClass, final String sql, Object... parameters) {
+    return executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters,
+        resultSet -> {
+          T ret = null;
+          if (resultSet.next()) {
+            ret = mapRowAux(objectClass, resultSet);
+          }
+          if (resultSet.next()) {
+            throw new SormException("Non-unique result returned");
+          }
+          return ret;
+        });
+  }
+
   @Override
-  public <T> T readOne(Class<T> objectClass, SqlStatement sql) {
-    return readOneAux(objectClass, sql.getSql(), sql.getParameters());
+  public <T1, T2, T3> List<Tuple3<T1, T2, T3>> readTupleList(Class<T1> t1, Class<T2> t2,
+      Class<T3> t3, SqlStatement sql) {
+    return readTupleList(t1, t2, t3, sql.getSql(), sql.getParameters());
+  }
+
+
+  @Override
+  public <T1, T2, T3> List<Tuple3<T1, T2, T3>> readTupleList(Class<T1> t1, Class<T2> t2,
+      Class<T3> t3, String sql, Object... parameters) {
+    List<Tuple3<T1, T2, T3>> ret =
+        executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters, resultSet -> {
+          final List<Tuple3<T1, T2, T3>> ret1 = new ArrayList<>();
+          while (resultSet.next()) {
+            ret1.add(
+                Tuples.of(getAux(t1, resultSet), getAux(t2, resultSet), getAux(t3, resultSet)));
+          }
+          return ret1;
+        });
+    return ret;
+  }
+
+  @Override
+  public <T1, T2> List<Tuple2<T1, T2>> readTupleList(Class<T1> t1, Class<T2> t2, SqlStatement sql) {
+    return readTupleList(t1, t2, sql.getSql(), sql.getParameters());
+  }
+
+  @Override
+  public <T1, T2> List<Tuple2<T1, T2>> readTupleList(Class<T1> t1, Class<T2> t2, String sql,
+      Object... parameters) {
+    List<Tuple2<T1, T2>> ret =
+        executeQueryAndRead(getJdbcConnection(), sqlParameterSetter, sql, parameters, resultSet -> {
+          final List<Tuple2<T1, T2>> ret1 = new ArrayList<>();
+          while (resultSet.next()) {
+            ret1.add(Tuples.of(getAux(t1, resultSet), getAux(t2, resultSet)));
+          }
+          return ret1;
+        });
+    return ret;
+  }
+
+  @Override
+  public <T> int[] update(List<T> objects) {
+    return applytoArray(objects, array -> update(array));
+  }
+
+  @Override
+  public <T> int update(T object) {
+    return getCastedTableMapping(object.getClass()).update(getJdbcConnection(), object);
+  }
+
+  @Override
+  public <T> int[] update(@SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(objects,
+        mapping -> mapping.update(getJdbcConnection(), objects), () -> new int[0]);
+  }
+
+  @Override
+  public <T> int[] updateOn(String tableName, List<T> objects) {
+    return applytoArray(objects, array -> updateOn(tableName, array));
+  }
+
+
+  @Override
+  public <T> int updateOn(String tableName, T object) {
+    return getCastedTableMapping(tableName, object.getClass()).update(getJdbcConnection(), object);
+  }
+
+
+  @Override
+  public <T> int[] updateOn(String tableName, @SuppressWarnings("unchecked") T... objects) {
+    return execSqlIfParameterExists(tableName, objects,
+        mapping -> mapping.update(getJdbcConnection(), objects), () -> new int[0]);
+  }
+
+  @Override
+  public void begin() {
+    begin(getTransactionIsolationLevel());
+  }
+
+  @Override
+  public void begin(int isolationLevel) {
+    setAutoCommit(false);
+    setTransactionIsolation(isolationLevel);
+  }
+
+  @Override
+  public void commit() {
+    Try.runOrThrow(() -> getJdbcConnection().commit(), Try::rethrow);
+  }
+
+  @Override
+  public void rollback() {
+    Try.runOrThrow(() -> getJdbcConnection().rollback(), Try::rethrow);
+  }
+
+  @Override
+  public void setAutoCommit(final boolean autoCommit) {
+    Try.runOrThrow(() -> getJdbcConnection().setAutoCommit(autoCommit), Try::rethrow);
+  }
+
+  private void setTransactionIsolation(int isolationLevel) {
+    Try.runOrThrow(() -> getJdbcConnection().setTransactionIsolation(isolationLevel), Try::rethrow);
   }
 
 }
